+++ Sources/StreamChatSwiftUI/ViewFactory/DefaultViewFactory.swift
@@ -84,7 +84,7 @@ extension ViewFactory {
     }
     
     public func makeChannelListBackground(options: ChannelListBackgroundOptions) -> some View {
-        Color(InjectedValues[\.colors].background)
+        Color(InjectedValues[\.colors].backgroundCoreApp)
             .edgesIgnoringSafeArea(.bottom)
     }
 
@@ -92,7 +92,7 @@ extension ViewFactory {
         options: ChannelListItemBackgroundOptions
     ) -> some View {
         let colors = InjectedValues[\.colors]
-        return Color(colors.backgroundElevation0)
+        return Color(colors.backgroundCoreElevation0)
     }
 
     public func makeChannelListDividerItem(options: ChannelListDividerItemOptions) -> some View {
@@ -193,7 +193,7 @@ extension ViewFactory {
     public func makeEmptyMessagesView(
         options: EmptyMessagesViewOptions
     ) -> some View {
-        Color(InjectedValues[\.colors].background)
+        Color(InjectedValues[\.colors].backgroundCoreApp)
             .frame(maxWidth: .infinity, maxHeight: .infinity)
             .accessibilityIdentifier("EmptyMessagesView")
     }
@@ -290,10 +290,11 @@ extension ViewFactory {
     public func makeMessageAttachmentsView(
         options: MessageAttachmentsViewOptions
     ) -> some View {
-        MessageAttachmentsView(
+        @Injected(\.utils) var utils
+        return MessageAttachmentsView(
             factory: self,
             message: options.message,
-            width: options.availableWidth,
+            width: min(options.availableWidth, utils.messageListConfig.attachmentPreviewWidth),
             isFirst: options.isFirst,
             scrolledId: options.scrolledId
         )
@@ -313,10 +314,11 @@ extension ViewFactory {
     public func makeGiphyAttachmentView(
         options: GiphyAttachmentViewOptions
     ) -> some View {
-        GiphyAttachmentView(
+        @Injected(\.utils) var utils
+        return GiphyAttachmentView(
             factory: self,
             message: options.message,
-            width: options.availableWidth,
+            width: min(options.availableWidth, utils.messageListConfig.attachmentPreviewWidth),
             isFirst: options.isFirst,
             scrolledId: options.scrolledId
         )
@@ -325,10 +327,11 @@ extension ViewFactory {
     public func makeLinkAttachmentView(
         options: LinkAttachmentViewOptions
     ) -> some View {
-        LinkAttachmentContainer(
+        @Injected(\.utils) var utils
+        return LinkAttachmentContainer(
             factory: self,
             message: options.message,
-            width: options.availableWidth,
+            width: min(options.availableWidth, utils.messageListConfig.attachmentPreviewWidth),
             isFirst: options.isFirst,
             scrolledId: options.scrolledId
         )
@@ -647,10 +650,11 @@ extension ViewFactory {
     public func makeVoiceRecordingView(
         options: VoiceRecordingViewOptions
     ) -> some View {
-        VoiceRecordingContainerView(
+        @Injected(\.utils) var utils
+        return VoiceRecordingContainerView(
             factory: self,
             message: options.message,
-            width: options.availableWidth,
+            width: min(options.availableWidth, utils.messageListConfig.attachmentPreviewWidth),
             isFirst: options.isFirst,
             scrolledId: options.scrolledId
         )
@@ -906,7 +910,14 @@ extension ViewFactory {
     }
     
     public func makePollView(options: PollViewOptions) -> some View {
-        PollAttachmentView(factory: self, message: options.message, poll: options.poll, isFirst: options.isFirst)
+        @Injected(\.utils) var utils
+        return PollAttachmentView(
+            factory: self,
+            message: options.message,
+            poll: options.poll,
+            isFirst: options.isFirst,
+            width: min(options.availableWidth, utils.messageListConfig.attachmentPreviewWidth)
+        )
     }
 
     // MARK: Threads
@@ -957,7 +968,7 @@ extension ViewFactory {
     }
 
     public func makeThreadListBackground(options: ThreadListBackgroundOptions) -> some View {
-        Color(options.colors.backgroundElevation1)
+        Color(options.colors.backgroundCoreElevation1)
             .edgesIgnoringSafeArea(.bottom)
     }
 
@@ -965,7 +976,7 @@ extension ViewFactory {
         options: ThreadListItemBackgroundOptions
     ) -> some View {
         let colors = InjectedValues[\.colors]
-        return Color(colors.backgroundElevation1)
+        return Color(colors.backgroundCoreElevation1)
     }
 
     public func makeThreadListDividerItem(options: ThreadListDividerItemOptions) -> some View {
@@ -981,7 +992,11 @@ extension ViewFactory {
     public func makeAttachmentTextView(
         options: AttachmentTextViewOptions
     ) -> some View {
-        AttachmentTextView(factory: self, message: options.message)
+        AttachmentTextView(
+            factory: self,
+            message: options.message,
+            availableWidth: options.availableWidth
+        )
     }
 
     public func makeStreamTextView(